refactor: relocate PageHeader component from SportteremContent to the sportterem igenyles page and update parseFormattedText import.
