+++ src/app/(app)/[lang]/sport/sportterem-igenyles/page.tsx
@@ -1,7 +1,7 @@
 import { getDictionary } from '@/get-dictionary';
 import type { Locale } from '@/i18n-config';
-import { PageHeader } from '@/components/common/PageHeader';
 import SportteremContent from './components/SportteremContent';
+import { PageHeader } from '@/components/common/PageHeader';
 type SportteremIgenylesPageProps = {
   params: Promise<{ lang: Locale }>;
 };
@@ -13,11 +13,10 @@ export default async function SportteremIgenylesPage({
   return (
     <div className="min-h-screen bg-gray-50">
       <div className="container mx-auto px-4 py-8">
-        <main className="container mx-auto py-10 px-4">
-          <SportteremContent 
-            content={dictionary.sport.sportterem} 
-          />
-        </main>
+        <PageHeader title={dictionary.sport.sportterem.title} />
+        <SportteremContent 
+          content={dictionary.sport.sportterem} 
+        />
       </div>
     </div>
   );
